[test] Use with_both_eh_sjlj in test_other.py. NFC
Followup to #21554
